Paciente Controller

+++ src/main/java/com/politecnico/sistema_citas_medicas/service/PacienteService.java
@@ -17,23 +17,47 @@ public class PacienteService  {
     @Autowired
     private final PacienteRepository pacienteRepository;
 
+
     public List<Paciente> obtenerPacientes() {
+
         return pacienteRepository.findAll();
     }
 
-    public Optional<Paciente> obtenerPacientePorId(Long id) {
-        return pacienteRepository.findById(id);
+    public Paciente obtenerPacientePorId(Long id) {
+
+        return pacienteRepository.findById(id).orElse(null);
     }
 
-    public Paciente guardaraPaciente(Paciente paciente) {
-        return pacienteRepository.save(paciente);
+    public Paciente obtenerPacientePorDocumento (String documento){
+
+        Paciente paciente = pacienteRepository.findByDocumento(documento);
+
+        //paciente 1000320709
+        if (paciente == null){
+            throw new IllegalArgumentException("Error: No existe un paciente con ese documento" + documento);
+        }
+        return paciente;
+    }
+
+    public Paciente guardarPaciente(Paciente paciente) {
+
+        Paciente pacienteExistente = pacienteRepository.findByDocumento(paciente.getDocumento());
+
+        if (pacienteExistente != null){
+
+            throw new IllegalArgumentException("Error: Ya existe un cliente registrado con el documento " + pacienteExistente.getDocumento());
+        } else {
+
+            return pacienteRepository.save(paciente);
+        }
     }
 
     public void eliminarPaciente(Long id) {
-        pacienteRepository.deleteById(id);
+
+        if (pacienteRepository.existsById(id)){
+            pacienteRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("No existe un paciente con ese id " + id);
+        }
     }
-
-
-
-
 }
